build: Move architecture specific code to arch/
