+++ kernel/arch/x86/gdt.asm
