+++ kernel/arch/x86/include/serial.h
@@ -1,7 +1,7 @@
 #ifndef SERIAL_H_FT2N9HWC
 #define SERIAL_H_FT2N9HWC
 
-#include "config.h"
+#include <config.h>
 
 #define SERIAL_PORT1 0x3f8
 #define SERIAL_PORT2 0x2f8
@@ -9,7 +9,7 @@
 #define SERIAL_PORT4 0x2e8
 
 #ifdef DEBUG_TO_SERIAL
-# include "io/serial/serial.h"
+# include <arch/x86/include/serial.h>
 # define DEBUG_SERIAL_PORT SERIAL_PORT1
 #endif
