+++ kernel/arch/x86/include/video.h
