+++ kernel/include/types.h
@@ -4,7 +4,7 @@
 #ifndef TYPES_H_VMW2LPX1
 #define TYPES_H_VMW2LPX1
 
-#include "include/asm/types.h"
+#include <include/asm/types.h>
 
 
 #ifndef _SIZE_T
